refactor: Rename file to match class
