+++ pet_mk_vii_planner/src/main.cpp
@@ -6,8 +6,8 @@
 #include "pet_mk_vii_planner/rrtDefinitions.hpp"
 #include "pet_mk_vii_planner/steerBezier.hpp"
 #include "pet_mk_vii_planner/steerCtrv.hpp"
+#include "rviz_visualizer.hpp"
 #include "utility/tiktok.hpp"
-#include "visualization.hpp"
 
 #include <ugl/lie_group/pose.h>
